Big Update

+++ Project1/Entity.hpp
@@ -1,6 +1,7 @@
 #ifndef ENTITY_HPP
 #define ENTITY_HPP
 
+#include <iostream>
 #include <SFML/Graphics.hpp>
 #include "Grid.hpp"
 
@@ -8,9 +9,9 @@ class Entity {
 public:
     sf::RectangleShape shape;
     sf::Vector2f velocity;
-    sf::Vector2f position;
-    Entity(sf::Vector2f position, sf::Color color);
-    virtual void update(float deltaTime, Grid& grid, sf::Vector2f playerPosition) = 0;
+    sf::Vector2i position;
+    Entity(sf::Vector2i position, sf::Color color);
+    virtual void update(float deltaTime, Grid& grid, sf::Vector2i playerPosition) = 0;
 };
 
 #endif // ENTITY_HPP
