+++ Project1/Player.hpp
@@ -6,8 +6,8 @@
 class Player : public Entity {
 public:
     static constexpr float SPEED = 200.0f;
-    Player(sf::Vector2f position);
-    void update(float deltaTime, Grid& grid, sf::Vector2f playerPosition) override;
+    Player(sf::Vector2i position);
+    void update(float deltaTime, Grid& grid, sf::Vector2i playerPosition) override;
 };
 
 #endif // PLAYER_HPP
